+++ src/main/java/pl/akademia/api/bike/Bike.java
@@ -1,4 +1,4 @@
-package pl.akademia.api.model;
+package pl.akademia.api.bike;
 
 import lombok.Data;
