+++ src/main/java/pl/akademia/api/bike/BikeRepository.java
@@ -1,12 +1,10 @@
-package pl.akademia.api;
+package pl.akademia.api.bike;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import pl.akademia.api.model.Bike;
 
-import javax.transaction.Transactional;
 import java.math.BigDecimal;
 
 @Repository
