+++ src/main/java/pl/akademia/api/bike/BikeRestController.java
@@ -1,4 +1,4 @@
-package pl.akademia.api;
+package pl.akademia.api.bike;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.akademia.api.model.Bike;
 
 import java.math.BigDecimal;
 import java.util.List;
