+++ src/main/java/pl/akademia/api/client/ClientRepository.java
@@ -3,7 +3,6 @@ package pl.akademia.api.client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import pl.akademia.api.model.Client;
 
 @Repository
 public interface ClientRepository extends JpaRepository<Client,Long> {
