+++ src/main/java/pl/akademia/api/client/ClientRestController.java
@@ -1,10 +1,8 @@
 package pl.akademia.api.client;
 
 
-import org.apache.logging.log4j.util.Strings;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,10 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import pl.akademia.api.model.Client;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 @RestController
