+++ src/main/java/pl/akademia/api/promotion/PromoCode.java
@@ -1,4 +1,4 @@
-package pl.akademia.api.model;
+package pl.akademia.api.promotion;
 
 
 import lombok.Data;
