+++ src/main/java/pl/akademia/api/promotion/PromoCodeRepository.java
@@ -1,9 +1,8 @@
-package pl.akademia.api;
+package pl.akademia.api.promotion;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import pl.akademia.api.model.PromoCode;
 
 import java.time.LocalDate;
 import java.util.List;
