+++ src/main/java/pl/akademia/api/user/UserApp.java
@@ -1,4 +1,4 @@
-package pl.akademia.api.model;
+package pl.akademia.api.user;
 
 import lombok.Data;
