+++ src/main/java/pl/akademia/api/user/UserRepository.java
@@ -1,7 +1,6 @@
-package pl.akademia.api;
+package pl.akademia.api.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import pl.akademia.api.model.UserApp;
 
 public interface UserRepository extends JpaRepository<UserApp, Long> {
 }
