+++ src/main/java/pl/akademia/api/user/UserRestController.java
@@ -1,9 +1,8 @@
-package pl.akademia.api;
+package pl.akademia.api.user;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.akademia.api.model.UserApp;
 
 import java.util.List;
