+++ src/main/java/pl/akademia/api/user/UserRole.java
@@ -1,4 +1,4 @@
-package pl.akademia.api.model;
+package pl.akademia.api.user;
 
 public enum UserRole {
