+++ src/main/java/pl/akademia/api/utils/DTOMapper.java
@@ -1,0 +1,5 @@
+package pl.akademia.api.utils;
+
+public interface DTOMapper<F,T>{
+  T from(F from);
+}
